Integrate TestThread
